backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
asyncpg==0.29.0
psycopg2-binary==2.9.9
weasyprint==62.3
num2words==0.5.13
anthropic==0.39.0
openai==1.51.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic-settings==2.5.2
pyjwt==2.9.0
bcrypt==4.2.0
stripe==10.12.0
resend==0.8.0
python-docx==1.1.0
markdown